+++ packages/core/legos/src/sources.ts
@@ -34,6 +34,7 @@ import * as coinmetrics from '@chainlink/coinmetrics-adapter'
 import * as coinpaprika from '@chainlink/coinpaprika-adapter'
 import * as coinranking from '@chainlink/coinranking-adapter'
 import * as covid_tracker from '@chainlink/covid-tracker-adapter'
+import * as cryptex from '@chainlink/cryptex-adapter'
 import * as cryptoapis from '@chainlink/cryptoapis-adapter'
 import * as cryptoapis_v2 from '@chainlink/cryptoapis-v2-adapter'
 import * as cryptocompare from '@chainlink/cryptocompare-adapter'
@@ -69,6 +70,7 @@ import * as geodb from '@chainlink/geodb-adapter'
 import * as google_bigquery from '@chainlink/google-bigquery-adapter'
 import * as gramchain from '@chainlink/gramchain-adapter'
 import * as graphql from '@chainlink/graphql-adapter'
+import * as gsr from '@chainlink/gsr-adapter'
 import * as iex_adapter from '@chainlink/iex-cloud-adapter'
 import * as intrinio from '@chainlink/intrinio-adapter'
 import * as ipfs from '@chainlink/ipfs-adapter'
@@ -77,6 +79,7 @@ import * as json_adapter from '@chainlink/json-rpc-adapter'
 import * as kaiko from '@chainlink/kaiko-adapter'
 import * as layer2_sequencer_health from '@chainlink/layer2-sequencer-health-adapter'
 import * as lcx from '@chainlink/lcx-adapter'
+import * as lido from '@chainlink/lido-adapter'
 import * as linkpool from '@chainlink/linkpool-adapter'
 import * as lition from '@chainlink/lition-adapter'
 import * as lotus from '@chainlink/lotus-adapter'
@@ -96,6 +99,7 @@ import * as paxos from '@chainlink/paxos-adapter'
 import * as paypal from '@chainlink/paypal-adapter'
 import * as poa from '@chainlink/poa-adapter'
 import * as polygon from '@chainlink/polygon-adapter'
+import * as por_address_list from '@chainlink/por-address-list-adapter'
 import * as reduce from '@chainlink/reduce-adapter'
 import * as renvm_adapter_address from '@chainlink/renvm-address-set-adapter'
 import * as satoshitango from '@chainlink/satoshitango-adapter'
@@ -162,6 +166,7 @@ export default {
   coinpaprika,
   coinranking,
   covid_tracker,
+  cryptex,
   cryptoapis,
   cryptoapis_v2,
   cryptocompare,
@@ -197,6 +202,7 @@ export default {
   google_bigquery,
   gramchain,
   graphql,
+  gsr,
   iex_adapter,
   intrinio,
   ipfs,
@@ -205,6 +211,7 @@ export default {
   kaiko,
   layer2_sequencer_health,
   lcx,
+  lido,
   linkpool,
   lition,
   lotus,
@@ -224,6 +231,7 @@ export default {
   paypal,
   poa,
   polygon,
+  por_address_list,
   reduce,
   renvm_adapter_address,
   satoshitango,