+++ packages/composites/google-weather/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @chainlink/google-weather-adapter
 
+## 1.0.53
+
+### Patch Changes
+
+- Updated dependencies [66396e888]
+- Updated dependencies [307aa10ec]
+  - @chainlink/ea-bootstrap@2.6.0
+  - @chainlink/google-bigquery-adapter@1.1.37
+
 ## 1.0.52
 
 ### Patch Changes