+++ packages/sources/amberdata/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @chainlink/amberdata-adapter
 
+## 1.4.8
+
+### Patch Changes
+
+- Updated dependencies [66396e888]
+- Updated dependencies [307aa10ec]
+  - @chainlink/ea-bootstrap@2.6.0
+  - @chainlink/ea-factories@1.0.51
+  - @chainlink/ea-test-helpers@1.2.0
+
 ## 1.4.7
 
 ### Patch Changes