+++ packages/sources/wbtc-address-set/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @chainlink/wbtc-address-set-adapter
 
+## 1.3.38
+
+### Patch Changes
+
+- Updated dependencies [66396e888]
+- Updated dependencies [307aa10ec]
+  - @chainlink/ea-bootstrap@2.6.0
+  - @chainlink/proof-of-reserves-adapter@1.5.18
+  - @chainlink/ea-test-helpers@1.2.0
+
 ## 1.3.37
 
 ### Patch Changes